backend/requirements.txt
fastapi
python-dotenv
pydantic
livekit
livekit-agents
livekit-api
livekit-plugins-deepgram
livekit-plugins-google
livekit-plugins-cartesia
livekit-plugins-silero
livekit-plugins-turn-detector
google-generativeai
